`ssh.ShellCommand` -> `command.WrapInLoginShell`
