+++ internal/install/install.go
@@ -35,7 +35,7 @@ type PathCandidate struct {
 
 func getPathDirs(targetDest ssh.Destination) ([]string, error) {
 	conn := target.NewConnection(targetDest, target.ConnectionOptions{})
-	output, err := conn.Run(ssh.ShellCommand("echo $PATH"))
+	output, err := conn.Run(command.WrapInLoginShell("echo $PATH"))
 	if err != nil {
 		return nil, err
 	}
@@ -48,7 +48,7 @@ func getPathDirs(targetDest ssh.Destination) ([]string, error) {
 
 func getHomeDir(targetDest ssh.Destination) (string, error) {
 	conn := target.NewConnection(targetDest, target.ConnectionOptions{})
-	output, err := conn.Run(ssh.ShellCommand("echo $HOME"))
+	output, err := conn.Run(command.WrapInLoginShell("echo $HOME"))
 	if err != nil {
 		return "", err
 	}
@@ -57,7 +57,7 @@ func getHomeDir(targetDest ssh.Destination) (string, error) {
 
 func getExistingBinaryDir(targetDest ssh.Destination, binaryName string) (string, error) {
 	conn := target.NewConnection(targetDest, target.ConnectionOptions{})
-	output, err := conn.Run(ssh.ShellCommand(fmt.Sprintf("command -v %s", binaryName)))
+	output, err := conn.Run(command.WrapInLoginShell(fmt.Sprintf("command -v %s", binaryName)))
 	if err != nil {
 		return "", nil
 	}
@@ -282,7 +282,7 @@ func install(installPath string, targetDest ssh.Destination, binaries map[string
 
 		installCmd := fmt.Sprintf("install -D -m %s /dev/stdin %s/%s", mode, installPath, binaryName)
 		conn := target.NewConnection(targetDest, target.ConnectionOptions{WithStdin: binaryData})
-		output, err := conn.Run(ssh.ShellCommand(installCmd))
+		output, err := conn.Run(command.WrapInLoginShell(installCmd))
 		if err != nil {
 			if errors.Is(err, ssh.ErrSSH) {
 				return err
