+++ internal/setupkeys/pubkeytransfer/pub_key_transfer.go
@@ -5,6 +5,7 @@ import (
 	"io"
 	"os"
 
+	"github.com/arm/topo/internal/command"
 	"github.com/arm/topo/internal/ssh"
 	target "github.com/arm/topo/internal/target"
 )
@@ -49,7 +50,7 @@ func (kt *PubKeyTransfer) Run(outputWriter io.Writer) error {
 	}
 
 	conn := kt.buildTransferConnection(pubKey)
-	cmdOutput, err := conn.Run(ssh.ShellCommand(remoteAuthorizedKeysCommand))
+	cmdOutput, err := conn.Run(command.WrapInLoginShell(remoteAuthorizedKeysCommand))
 	if err != nil {
 		return fmt.Errorf("failed to transfer public key to target %s: %w", kt.dest, err)
 	}
@@ -59,7 +60,7 @@ func (kt *PubKeyTransfer) Run(outputWriter io.Writer) error {
 
 func (kt *PubKeyTransfer) DryRun(output io.Writer) error {
 	conn := kt.buildTransferConnection(nil)
-	if err := conn.DryRun(ssh.ShellCommand(remoteAuthorizedKeysCommand), output); err != nil {
+	if err := conn.DryRun(command.WrapInLoginShell(remoteAuthorizedKeysCommand), output); err != nil {
 		return fmt.Errorf("failed to write dry-run output for public key transfer to target %s: %w", kt.dest, err)
 	}
 	return nil
